requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pymongo==4.13.0
python-dotenv==1.1.0
requests==2.32.3